html/semantics/embedded-content/media-elements/user-interface/muted.html

<!doctype html>
<title>muted</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<style>video { display: none; }</style>
<div id=log></div>

<!-- These tests are inside <video> so that the steps for updating the
     muted IDL attribute cannot be delayed until </video>. -->

<video id=v1>
<script>
var v1 = document.getElementById('v1');

test(function() {
  assert_false(v1.muted);
}, 'getting video.muted (parser-created)');

test(function() {
  v1.muted = true;
  assert_true(v1.muted);
  assert_false(v1.hasAttribute('muted'));

  v1.muted = false;
  assert_false(v1.muted);
  assert_false(v1.hasAttribute('muted'));
}, 'setting video.muted (parser-created)');
</script>
</video>

<video id=v2 muted>
<script>
var v2 = document.getElementById('v2');

test(function() {
  assert_true(v2.muted);
}, 'getting video.muted with muted="" (parser-created)');

test(function() {
  v2.muted = false;
  assert_false(v2.muted);
  assert_true(v2.hasAttribute('muted'));

  v2.muted = true;
  assert_true(v2.muted);
  assert_true(v2.hasAttribute('muted'));
}, 'setting video.muted with muted="" (parser-created)');
</script>
</video>

<!-- Negative test to ensure that the load algorithm does not update the
     muted IDL attribute to match the content attribute. -->

<video id=v3 muted></video>
<script>
async_test(function(t) {
  var v = document.getElementById('v3');
  assert_true(v.muted);
  v.muted = false;
  v.src = 'data:,'; // invokes load()
  v.addEventListener('error', t.step_func(function() {
    assert_false(v.muted);
    t.done();
  }));
}, 'getting video.muted with muted="" after load (parser-created)');
</script>

<script>
test(function() {
  var v = document.createElement('video');
  assert_false(v.muted);
}, 'getting video.muted (script-created)');

test(function() {
  var v = document.createElement('video');
  v.setAttribute('muted', '');
  assert_false(v.muted);
}, 'getting video.muted with muted="" (script-created)');

test(function() {
  var v = document.createElement('video');
  v.setAttribute('muted', '');
  v = v.cloneNode(false);
  assert_true(v.hasAttribute('muted'));
  assert_false(v.muted);
}, 'getting video.muted with muted="" (cloneNode-created)');

test(function() {
  var e = document.createElement('div');
  e.innerHTML = '<video muted>';
  v = e.firstChild;
  assert_true(v.hasAttribute('muted'));
  assert_true(v.muted);
}, 'getting video.muted with muted="" (innerHTML-created)');
</script>
